+++ rust/chains/hyperlane-sealevel/src/mailbox.rs
@@ -113,7 +113,7 @@ impl SealevelMailbox {
 
     /// Simulates an instruction, and attempts to deserialize it into a T.
     /// If no return data at all was returned, returns Ok(None).
-    /// If some return data was returned but deserialization was unsuccesful,
+    /// If some return data was returned but deserialization was unsuccessful,
     /// an Err is returned.
     pub async fn simulate_instruction<T: BorshDeserialize + BorshSerialize>(
         &self,