+++ typescript/infra/config/environments/mainnet3/gas-oracle.ts
@@ -50,6 +50,7 @@ const gasPrices: ChainMap<BigNumber> = {
   polygonzkevm: ethers.utils.parseUnits('2', 'gwei'),
   neutron: ethers.utils.parseUnits('1', 'gwei'),
   mantapacific: ethers.utils.parseUnits('1', 'gwei'),
+  viction: ethers.utils.parseUnits('0.25', 'gwei'),
 };
 
 // Accurate from coingecko as of Mar 9, 2023.
@@ -92,6 +93,8 @@ const tokenUsdPrices: ChainMap<BigNumber> = {
     '1619.00',
     TOKEN_EXCHANGE_RATE_DECIMALS,
   ),
+  // https://www.coingecko.com/en/coins/viction
+  viction: ethers.utils.parseUnits('0.881', TOKEN_EXCHANGE_RATE_DECIMALS),
 };
 
 // Gets the exchange rate of the remote quoted in local tokens