+++ typescript/infra/scripts/agent-utils.ts
@@ -1,3 +1,4 @@
+import debug from 'debug';
 import path from 'path';
 import yargs from 'yargs';
 
@@ -29,6 +30,8 @@ import { fetchProvider } from '../src/config/chain';
 import { EnvironmentNames, deployEnvToSdkEnv } from '../src/config/environment';
 import { Role } from '../src/roles';
 import { assertContext, assertRole, readJSON } from '../src/utils/utils';
+
+const debugLog = debug('infra:scripts:utils');
 
 export enum Modules {
   // TODO: change
@@ -70,6 +73,13 @@ export function withModuleAndFork<T>(args: yargs.Argv<T>) {
     .describe('fork', 'network to fork')
     .choices('fork', Object.values(Chains))
     .alias('f', 'fork');
+}
+
+export function withNetwork<T>(args: yargs.Argv<T>) {
+  return args
+    .describe('network', 'network to target')
+    .choices('network', Object.values(Chains))
+    .alias('n', 'network');
 }
 
 export function withContext<T>(args: yargs.Argv<T>) {
@@ -231,6 +241,7 @@ export function getKeyForRole(
   role: Role,
   index?: number,
 ): CloudAgentKey {
+  debugLog(`Getting key for ${role} role`);
   const agentConfig = getAgentConfig(context, environment);
   return getCloudAgentKey(agentConfig, role, chain, index);
 }
@@ -244,7 +255,9 @@ export async function getMultiProviderForRole(
   // TODO: rename to consensusType?
   connectionType?: RpcConsensusType,
 ): Promise<MultiProvider> {
+  debugLog(`Getting multiprovider for ${role} role`);
   if (process.env.CI === 'true') {
+    debugLog('Returning multiprovider with default RPCs in CI');
     return new MultiProvider(); // use default RPCs
   }
   const multiProvider = new MultiProvider(txConfigs);
@@ -271,6 +284,7 @@ export async function getKeysForRole(
   index?: number,
 ): Promise<ChainMap<CloudAgentKey>> {
   if (process.env.CI === 'true') {
+    debugLog('No keys to return in CI');
     return {};
   }