+++ typescript/infra/scripts/gas-oracle/update-storage-gas-oracle.ts
@@ -4,7 +4,7 @@ import { ProtocolType } from '@hyperlane-xyz/utils';
 import { RemoteGasData, StorageGasOracleConfig } from '../../src/config';
 import { deployEnvToSdkEnv } from '../../src/config/environment';
 import { RemoteGasDataConfig } from '../../src/config/gas-oracle';
-import { getArgs } from '../agent-utils';
+import { getArgs, withNetwork } from '../agent-utils';
 import { getEnvironmentConfig } from '../core-utils';
 
 import {
@@ -20,12 +20,12 @@ import {
  * Expects the deployer key to be the owner of the StorageGasOracle contract.
  */
 async function main() {
-  const args = await getArgs()
+  const args = await withNetwork(getArgs())
     .boolean('dry-run')
     .describe('dry-run', 'If true, will not submit any transactions')
     .default('dry-run', false).argv;
 
-  const { environment } = await getArgs().argv;
+  const { environment, network } = args;
   const coreEnvConfig = getEnvironmentConfig(environment);
   const multiProvider = await coreEnvConfig.getMultiProvider();
 
@@ -39,7 +39,9 @@ async function main() {
     multiProvider,
   );
 
-  for (const chain of igp.chains()) {
+  const targetChains = network ? [network] : igp.chains();
+
+  for (const chain of targetChains) {
     if (
       multiProvider.getChainMetadata(chain).protocol !== ProtocolType.Ethereum
     ) {
@@ -72,7 +74,7 @@ async function setStorageGasOracleValues(
 
   const configsToSet: RemoteGasDataConfig[] = [];
 
-  for (const remote in localStorageGasOracleConfig) {
+  for (const remote of Object.keys(localStorageGasOracleConfig)) {
     const desiredGasData = localStorageGasOracleConfig[remote]!;
     const remoteId = multiProvider.getDomainId(remote);